requirements.txt
google-generativeai==0.8.5
PyMuPDF==1.25.5
streamlit==1.44.1
python-dotenv==1.1.0